ai/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ru-fin-ai-helper"
version = "0.1.0"
description = "Deterministic AI helper stubs for ru-fin-monitoring-helper"
readme = "../README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
